--- frontend/composables/recipes/use-recipe-context.ts
@@ -1,43 +1,0 @@
-import { useAsync, ref } from "@nuxtjs/composition-api";
-import { useUserApi } from "~/composables/api";
-import { Recipe } from "~/types/api-types/recipe";
-
-export const useRecipeContext = function () {
-  const api = useUserApi();
-  const loading = ref(false);
-
-  function getBySlug(slug: string) {
-    loading.value = true;
-    const recipe = useAsync(async () => {
-      const { data } = await api.recipes.getOne(slug);
-      return data;
-    }, slug);
-
-    loading.value = false;
-
-    return recipe;
-  }
-
-  async function fetchRecipe(slug: string) {
-    loading.value = true;
-    const { data } = await api.recipes.getOne(slug);
-    loading.value = false;
-    return data;
-  }
-
-  async function deleteRecipe(slug: string) {
-    loading.value = true;
-    const { data } = await api.recipes.deleteOne(slug);
-    loading.value = false;
-    return data;
-  }
-
-  async function updateRecipe(slug: string, recipe: Recipe) {
-    loading.value = true;
-    const { data } = await api.recipes.updateOne(slug, recipe);
-    loading.value = false;
-    return data;
-  }
-
-  return { loading, getBySlug, deleteRecipe, updateRecipe, fetchRecipe };
-};
