+++ frontend/pages/recipe/_slug/ingredient-parser.vue
@@ -76,7 +76,7 @@ import { defineComponent, ref, useRoute, useRouter } from "@nuxtjs/composition-a
 import { Food, ParsedIngredient, Parser } from "~/api/class-interfaces/recipes";
 import RecipeIngredientEditor from "~/components/Domain/Recipe/RecipeIngredientEditor.vue";
 import { useUserApi } from "~/composables/api";
-import { useRecipeContext, useFoods, useUnits } from "~/composables/recipes";
+import { useRecipe, useFoods, useUnits } from "~/composables/recipes";
 import { RecipeIngredientUnit } from "~/types/api-types/recipe";
 
 interface Error {
@@ -99,9 +99,7 @@ export default defineComponent({
     const slug = route.value.params.slug;
     const api = useUserApi();
 
-    const { getBySlug, loading } = useRecipeContext();
-
-    const recipe = getBySlug(slug);
+    const { recipe, loading } = useRecipe(slug);
 
     const ingredients = ref<any[]>([]);
